UI/ like, comment integration

+++ back-end/LowLand/src/main/java/com/coffee/lowland/DTO/request/order/CancelPaymentRequest.java
@@ -13,6 +13,6 @@ import lombok.experimental.FieldDefaults;
 @Builder
 public class CancelPaymentRequest {
     @NotNull(message = "ORDER_FIELD_NOT_BE_BLANK")
-    int orderCode;
+    long orderCode;
     String reason;
 }
