+++ back-end/LowLand/src/main/java/com/coffee/lowland/model/Order.java
@@ -2,17 +2,11 @@ package com.coffee.lowland.model;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 
 import java.time.LocalDateTime;
-import java.time.ZonedDateTime;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-import java.util.UUID;
 
 @Entity
 @Getter
@@ -24,7 +18,7 @@ public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
     String orderId;
-    int orderCode;
+    long orderCode;
     String accountId;
     String customerName;
     String phoneNumber;
